- comito

+++ 03_PythonList/main.cpp
@@ -3,7 +3,7 @@
 
 int main()
 {
-    list<int> a; // float has problems for some reason
+    list<float> a;
     a.append(1);
     a.append(2);
     a.append(3);
